+++ gitlab-harvester-operator/src/js/main.js
@@ -173,6 +173,10 @@
                 for (var i = 0; i < data.length; i++) {
                     commits.push(normalizeCommit(data[i]));
                 }
+                // Revert the order so newer commits are last
+                // This is for the graphs to sow first on the left and latest on the right
+                commits.reverse();
+
                 //Add some metadata
                 commits.metadata = {};
                 commits.metadata.type = "list";
@@ -207,6 +211,10 @@
         for (var i = 0; i < data.length; i++) {
             result.push(normalizeIssue(data[i]));
         }
+
+        // Revert the order so newer issues are last
+        // This is for the graphs to sow first on the left and latest on the right
+        result.reverse();
 
         return result;
     };